# swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Asset API MongoDB with OAuth",
      "description": "[NOTE: The W10 working Endpoint is the first one listed:  '/assets/']. CSE 341 Final Project Asset API MongoDB with OAuth. All routes will be protected, and /create, /update, and /delete require specific permissions.",
      "version": "1.0"
    },
    "host": "asset-tracker-ws7z.onrender.com",
    "basePath": "/",
    "schemes": [
      "https"
    ],
    "paths": {
...
        "/assets/{assetId}": {
          "get": {
            "tags": [
                "Assets"
            ],
            "description": "Get single asset by id (e.g., 'AS010'). Asset ID must be two alpha characters (e.g., 'AS') that are case-insensitive, followed by three digits, padded left with zeros (e.g., '005').",
            "example": "'AS010', 'AS004', 'AS012'",
            "parameters": [
              {
                "name": "assetId",
                "in": "path",
                "required": true,
                "type": "string"
              }
            ],
            "responses": {
              "200": {
                "description": "OK"
              }
            }
          },
          "put": {
            "tags": [
                "Assets"
            ],
            "description": "Update an asset by id (e.g., 'AS010'). Asset ID must be two alpha characters (e.g., 'AS') that are case-insensitive, followed by three digits, padded left with zeros (e.g., '005').",
            "example": "'AS010', 'AS004', 'AS012'",
            "parameters": [
              {
                "name": "assetId",
                "in": "path",
                "required": true,
                "type": "string"
              },
              {
                "name": "body",
                "in": "body",
                "schema": {
                  "type": "object",
                  "example": {
                    "SerialNumber": "EQ-934R2",
                    "Brand": "Sony",
                    "PurchaseDate": "18 Dec 2019",
                    "Model": "Vista",
                    "ModelNum": "VA-934",
                    "PurchasePrice": "13900",
                    "image": "https://images-na.ssl-images-amazon.com/images/I/71%2B%2B%2B%2Bq%2BqgL._AC_SL1500_.jpg",
                    "PhysicalDescription": "4k display, 32GB RAM, 1TB SSD",
                    "Status": "active",
                    "Condition": "good",
                    "Building": "West 3B",
                    "User": "rsmithson@tesla.com"
                  }
                }
              }
            ],
            "responses": {
              "204": {
                "description": "No Content"
              },
              "201": {
                "description": "Record updated successfully"
              },
              "500": {
                "description": "Internal Server Error"
              }
            }
          },
          "delete": {
            "tags": [
                "Assets"
            ],
            "description": "Delete an asset by id (e.g., 'AS010'). Asset ID must be two alpha characters (e.g., 'AS') that are case-insensitive, followed by three digits ('005').",
            "example": "'AS010', 'AS004', 'AS012'",
            "parameters": [
              {
                "name": "assetId",
                "in": "path",
                "required": true,
                "type": "string"
              }
            ],
            "responses": {
              "200": {
                "description": "Record successfully deleted"
              },
              "500": {
                "description": "Record deletion unsuccessful. Internal Server Error"
              },
              "404": {
                "description": "Deletion unsuccessful, record Not Found"
              }
            }
          }
        },
        "/assets": {
          "post": {
            "tags": [
                "Assets"
            ],
            "description": "Create an asset. Asset ID must be two alpha characters (e.g., 'AS') that are case-insensitive, followed by three digits, padded left with zeros (e.g., '005').",
            "parameters": [
              {
                "name": "body",
                "in": "body",
                "schema": {
                  "type": "object",
                  "example": {
                    "AssetId": "AS-004",
                    "SerialNumber": "EQ-934R2",
                    "Brand": "Sony",
                    "PurchaseDate": "18 Dec 2019",
                    "Model": "Vista",
                    "ModelNum": "VA-934",
                    "PurchasePrice": "13900",
                    "image": "https://images-na.ssl-images-amazon.com/images/I/71%2B%2B%2B%2Bq%2BqgL._AC_SL1500_.jpg",
                    "PhysicalDescription": "4k display, 32GB RAM, 1TB SSD",
                    "Status": "active",
                    "Condition": "good",
                    "Building": "West 3B",
                    "User": "rsmithson@tesla.com"
                  }
                }
              }
            ],
            "responses": {
              "201": {
                "description": "Record creation successful."
              },
              "400": {
                "description": "Record creation unsuccessful. Bad Request - Invalid input or duplicate record found"
              },
              "500": {
                "description": "Record creation unsuccessful. Internal Server Error"
              }
            }
          }
        },
        "/assets/param/{paramType}/{paramValue}": {
          "get": {
            "tags": [
                "Assets"
            ],
            "description": "Get assets by parameter type and value",
            "parameters": [
              {
                "name": "paramType",
                "in": "path",
                "required": true,
                "type": "string"
              },
              {
                "name": "paramValue",
                "in": "path",
                "required": true,
                "type": "string"
              }
            ],
            "responses": {
              "200": {
                "description": "OK"
              }
            }
          }
        },
...
    }
}